+++ example.js
@@ -1,3 +1,0 @@
-const createDiff = require("./index")
-
-createDiff("maps/v1.dat", "maps/v2.dat", "diff").then(diff => console.log(diff))